test_multiple_async_display is not really about async and it's not about multiple display. It is about implicit display() target, so move it it the correct class and improve the test to be less dependent on the exact <div> id
